Refactor the Event Handler directory to make it easier for us to work on many routes at the same time
